+++ test/test_distrib_pkg.py
@@ -96,6 +96,7 @@ def test_get_sig_url(has_sig: bool) -> None:
                     "data-gpg-sig": "true",
                     "data-core-metadata": "sha256=ae718719df4708f329d58ca4d5390c1206c4222ef7e62a3aa9844397c63de28b",
                     "data-yanked": "Oopsy.",
+                    "data-provenance": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
                 },
             ),
             DistributionPackage(
@@ -115,6 +116,7 @@ def test_get_sig_url(has_sig: bool) -> None:
                     "sha256": "ae718719df4708f329d58ca4d5390c1206c4222ef7e62a3aa9844397c63de28b"
                 },
                 has_metadata=True,
+                provenance_sha256="e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
             ),
         ),
         (
@@ -197,6 +199,27 @@ def test_pep658_no_digests() -> None:
     )
 
 
+def test_provenance_url() -> None:
+    pkg = DistributionPackage(
+        filename="qypi-0.1.0-py3-none-any.whl",
+        url="https://files.pythonhosted.org/packages/82/fc/9e25534641d7f63be93079bc07fa92bab136ddf5d4181059a1308a346f96/qypi-0.1.0-py3-none-any.whl",
+        digests={
+            "sha256": "da69d28dcd527c0e372b3fa7b92fc333b327f8470175f035abc4e351b539189f"
+        },
+        has_sig=True,
+        requires_python="~= 3.6",
+        project="qypi",
+        version="0.1.0",
+        package_type="wheel",
+        is_yanked=False,
+        yanked_reason=None,
+    )
+    assert (
+        pkg.provenance_url
+        == "https://files.pythonhosted.org/packages/82/fc/9e25534641d7f63be93079bc07fa92bab136ddf5d4181059a1308a346f96/qypi-0.1.0-py3-none-any.whl.provenance"
+    )
+
+
 def test_from_json_data_no_metadata() -> None:
     pkg = DistributionPackage.from_json_data(
         {
@@ -241,3 +264,22 @@ def test_from_json_data_metadata(
     )
     assert pkg.has_metadata == has_metadata
     assert pkg.metadata_digests == metadata_digests
+
+
+def test_from_json_data_provenance() -> None:
+    pkg = DistributionPackage.from_json_data(
+        {
+            "filename": "argset-0.1.0-py3-none-any.whl",
+            "hashes": {
+                "sha256": "107a632c7112faceb9fd6e93658dd461154713db250f7ffde5bd473e17cf1db5"
+            },
+            "requires-python": "~=3.6",
+            "url": "https://files.pythonhosted.org/packages/b5/2b/7aa284f345e37f955d86e4cd57b1039b573552b0fc29d1a522ec05c1ee41/argset-0.1.0-py3-none-any.whl",
+            "yanked": False,
+            "provenance": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+        }
+    )
+    assert (
+        pkg.provenance_sha256
+        == "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
+    )