+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 v1.6.0 (in development)
 -----------------------
 - Drop support for Python 3.7
+- Support PEP 740
+    - `provenance_sha256` and `provenance_url` fields added to
+      `DistributionPackage`
+    - `get_provenance()` method added to `PyPISimple`
+    - `NoProvenanceError` exception type added
 
 v1.5.0 (2024-02-24)
 -------------------